requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.3.0
bcrypt==4.0.1
bidict==0.23.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.7
frozenlist==1.5.0
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.6.1
multidict==6.1.0
orjson==3.10.14
passlib==1.7.4
pillow==11.0.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.0
pydantic-extra-types==2.10.1
pydantic-settings==2.7.1
pydantic_core==2.27.0
Pygments==2.19.1
PyJWT==2.7.0
pymongo==4.9.2
python-dotenv==1.0.1
python-engineio==4.11.2
python-multipart==0.0.17
python-socketio==5.12.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
resend==2.4.0
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
shellingham==1.5.4
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.41.3
twilio==9.4.3
typer==0.15.1
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.34.0
watchfiles==1.0.4
websockets==14.1
wsproto==1.2.0
yarl==1.18.3
